refactor(load): move load test interfaces to a separate file and improve schema deletion
- Moves `LoadTestConfig`, `SubmissionResult`, and `TestMetrics` interfaces from `load-test.service.ts` to a new file `load-test.types.ts` to improve code organization and reusability.
- Updates imports in `load-test.service.ts` and `submit-answer.load-spec.ts` to reflect the new location of the interfaces.
- Fixes a type issue in `setup-load.ts` by explicitly defining the type of the raw SQL execution result as `<string>`.
