+++ packages/challenges/test/setup-load.ts
@@ -43,6 +43,6 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-  await prisma.$executeRawUnsafe(`DROP SCHEMA IF EXISTS "${schemaId}" CASCADE`);
+  await prisma.$executeRawUnsafe<string>(`DROP SCHEMA IF EXISTS "${schemaId}" CASCADE`);
   await prisma.$disconnect();
 });
